--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "ramls/raml-util"]
-	path = ramls/raml-util
-	url = https://github.com/folio-org/raml
