+++ .rancher-pipeline.yml
@@ -1,0 +1,27 @@
+stages:
+- name: Build
+  steps:
+  - runScriptConfig:
+      image: maven:3-openjdk-11
+      shellScript: mvn package -DskipTests
+- name: Build Docker with DIND
+  steps:
+  - publishImageConfig:
+      dockerfilePath: ./Dockerfile
+      buildContext: .
+      tag: docker.dev.folio.org/mod-data-export-spring:scout-latest
+      pushRemote: true
+      registry: docker.dev.folio.org
+- name: Deploy
+  steps:
+  - applyAppConfig:
+      catalogTemplate: p-ngvqv:scout-helmcharts-mod-data-export-spring
+      version: 0.1.26
+      answers:
+        image.repository: docker.dev.folio.org/mod-data-export-spring
+        image.tag: scout-latest
+        postJob.enabled: true
+      targetNamespace: scout
+      name: mod-data-export-spring
+timeout: 60
+notification: {}
