--- src/main/java/org/folio/DataExportApplication.java
@@ -1,12 +1,0 @@
-package org.folio;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DataExportApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DataExportApplication.class, args);
-    }
-}
