+++ src/main/java/org/folio/des/ModDataExportSpringApplication.java
@@ -1,0 +1,13 @@
+package org.folio.des;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ModDataExportSpringApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(ModDataExportSpringApplication.class, args);
+  }
+
+}
