+++ src/main/java/org/folio/des/domain/dto/BaseJobDto.java
@@ -1,0 +1,13 @@
+package org.folio.des.domain.dto;
+
+import lombok.Data;
+import org.folio.des.domain.entity.enums.JobType;
+
+@Data
+public class BaseJobDto {
+
+  private String name;
+  private String description;
+  private JobType jobType;
+
+}
