+++ src/main/java/org/folio/des/domain/dto/JobDto.java
@@ -1,0 +1,15 @@
+package org.folio.des.domain.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.util.UUID;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class JobDto extends BaseJobDto {
+
+  private UUID id;
+  private JobExecutionDto jobExecution;
+
+}
