+++ src/main/java/org/folio/des/domain/dto/JobExecutionUpdateDto.java
@@ -1,0 +1,14 @@
+package org.folio.des.domain.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.util.UUID;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class JobExecutionUpdateDto extends JobExecutionDto {
+
+  private UUID jobId;
+
+}
