+++ src/main/java/org/folio/des/domain/dto/StartJobCommandDto.java
@@ -1,0 +1,14 @@
+package org.folio.des.domain.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.util.UUID;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class StartJobCommandDto extends StartJobRequestDto {
+
+  private UUID id;
+
+}
