+++ src/main/java/org/folio/des/domain/dto/StartJobRequestDto.java
@@ -1,0 +1,14 @@
+package org.folio.des.domain.dto;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.util.Map;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class StartJobRequestDto extends BaseJobDto {
+
+  private Map<String, JobParameterDto> jobInputParameters;
+
+}
