+++ src/main/java/org/folio/des/domain/entity/enums/JobType.java
@@ -1,0 +1,5 @@
+package org.folio.des.domain.entity.enums;
+
+public enum JobType {
+  CIRCULATION_LOG_EXPORT, BURSAR_FEES_FINES_EXPORT
+}
