+++ src/main/java/org/folio/des/domain/entity/enums/ParameterType.java
@@ -1,0 +1,5 @@
+package org.folio.des.domain.entity.enums;
+
+public enum ParameterType {
+  STRING, DATE, LONG, DOUBLE
+}
