+++ src/main/java/org/folio/des/repository/IJobRepository.java
@@ -1,0 +1,16 @@
+package org.folio.des.repository;
+
+import org.folio.des.domain.entity.Job;
+import org.folio.des.domain.entity.JobExecution;
+
+import java.util.UUID;
+
+public interface IJobRepository {
+
+  Job createJob(Job job);
+
+  void updateJobExecution(UUID jobId, JobExecution jobExecution);
+
+  Job getJob(UUID jobId);
+
+}
