--- src/main/java/org/folio/dto/GreetingDto.java
@@ -1,13 +1,0 @@
-package org.folio.dto;
-
-import lombok.Data;
-
-@Data
-public class GreetingDto {
-
-    private int id;
-
-    private String greeting;
-
-    private String language;
-}
