--- src/main/java/org/folio/dto/JobExecutionUpdateDto.java
@@ -1,11 +1,0 @@
-package org.folio.dto;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class JobExecutionUpdateDto extends JobExecutionDto {
-
-    private UUID jobId;
-}
