--- src/main/java/org/folio/dto/StartJobCommandDto.java
@@ -1,11 +1,0 @@
-package org.folio.dto;
-
-import lombok.Data;
-
-import java.util.UUID;
-
-@Data
-public class StartJobCommandDto extends StartJobRequestDto {
-
-    private UUID id;
-}
