--- src/main/java/org/folio/dto/StartJobRequestDto.java
@@ -1,11 +1,0 @@
-package org.folio.dto;
-
-import lombok.Data;
-
-import java.util.Map;
-
-@Data
-public class StartJobRequestDto extends BaseJobDto {
-
-    private Map<String, JobParameterDto> jobInputParameters;
-}
