--- src/main/java/org/folio/model/entities/enums/JobType.java
@@ -1,5 +1,0 @@
-package org.folio.model.entities.enums;
-
-public enum JobType {
-    GREETINGS_EXPORT
-}
