--- src/main/java/org/folio/model/entities/enums/ParameterType.java
@@ -1,8 +1,0 @@
-package org.folio.model.entities.enums;
-
-public enum ParameterType {
-    STRING,
-    DATE,
-    LONG,
-    DOUBLE
-}
