+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
-server.port=8082
+server.port=8081
+spring.application.name=mod-data-export-spring
 
 spring.kafka.bootstrap-servers=localhost:9092
 spring.kafka.consumer.group-id=jobExecutionUpdateCG
@@ -8,7 +9,4 @@ spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.seria
 spring.kafka.producer.key-serializer=org.springframework.kafka.support.serializer.JsonSerializer
 spring.kafka.producer.value-serializer=org.springframework.kafka.support.serializer.JsonSerializer
 
-
-
-
-
+folio.tenant.validation.enabled=true
