+++ src/main/resources/swagger.api/examples/error.sample
@@ -1,0 +1,11 @@
+{
+  "message": "may not be null",
+  "type": "1",
+  "code": "-1",
+  "parameters": [
+    {
+      "key": "moduleTo",
+      "value": "null"
+    }
+  ]
+}
