+++ src/main/resources/swagger.api/schemas/common/metadata.json
@@ -1,0 +1,39 @@
+{
+  "title": "Metadata Schema",
+  "description": "Metadata about creation and changes to records, provided by the server (client should not provide)",
+  "type": "object",
+  "properties": {
+    "createdDate": {
+      "description": "Date and time when the record was created",
+      "type": "string",
+      "format": "date-time"
+    },
+    "createdByUserId": {
+      "description": "ID of the user who created the record (when available)",
+      "type": "string",
+      "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
+    },
+    "createdByUsername": {
+      "description": "Username of the user who created the record (when available)",
+      "type": "string"
+    },
+    "updatedDate": {
+      "description": "Date and time when the record was last updated",
+      "type": "string",
+      "format": "date-time"
+    },
+    "updatedByUserId": {
+      "description": "ID of the user who last updated the record (when available)",
+      "type": "string",
+      "pattern": "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
+    },
+    "updatedByUsername": {
+      "description": "Username of the user who last updated the record (when available)",
+      "type": "string"
+    }
+  },
+  "additionalProperties": false,
+  "required": [
+    "createdDate"
+  ]
+}
